requirements.txt
piper-tts
numpy
scipy
google-cloud-firestore
requests
